requirements.txt
einops==0.6.1
filelock==3.12.2
jsonlines==3.1.0
lmdb==1.4.1
matplotlib==3.5.1
msgpack_numpy==0.4.8
msgpack_python==0.5.6
numpy==1.23.5
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
Pillow==9.4.0
tensorboardX==2.6
torch==1.13.0
torchvision==0.14.0
tqdm==4.62.3
transformers==4.19.4
typed_argument_parser==1.8.0
yacs==0.1.8
absl-py==1.4.0
